Genericize namespace cache TTL logic for all resources
Contributes to #10.

Namespaces are cached with a TTL determined by the variable
`kele-namespace-refresh-interval`. This logic cleanly generalizes to all
resources. This commit does so. We also implement a mechanism by which users (or
the package author) can define a resource-specific TTL override -- useful for
resources that are expected to change only rarely, e.g. namespaces.