Print a progress reporter when loading kubeconfig changes
Now that kubeconfig loading is performed asynchronously, users are no longer
blocked on the loading process. The trade-off now is that users have no idea of
a) when kubeconfig loading is taking place (and the results of their other
`kele-` operations are potentially stale) and b) when it has completed (and
their `kele-` operations are now guaranteed to be up-to-date and accurate).

To remedy that opacity, we use a [progress reporter][1] to print a conventional
"doing thing..." message when the subroutine starts, and a "doing thing...done"
message when it concludes.

[1]: https://www.gnu.org/software/emacs/manual/html_node/elisp/Progress.html