Implement MVP of `kele-get` for getting + displaying resources (#70)
Closes #64.

This commit introduces an MVP version of `kele-get`, an interactive
function to fetch a specific resource and display its manifest in a
separate buffer.

In the name of an MVP, we note several sub-optimal user experiences:
- Invocation hangs on the synchronous proxy server instantiation
necessary for namespace completion. It seems feasible to start the proxy
server asynchronously and pass that future to the namespace completion
function, which would allow the proxy server to spin up while the user
is inputting the group-version, kind, etc. See #72 
- There is no completion of the resource NAME, i.e. user needs to
already know the name of the thing they're trying to GET. This is
troublesome w/, say, pods; anybody who's run `kubectl get pods`,
highlighted and copied a specific pod name, then run `kubectl get pods
<pod-name>` again can attest 🙃 See #73