Implement Transient prefix for acting on resources (#114)
Closes #109.

This commit introduces a rudimentary Transient prefix for working with
resources, `kele-resource`. Each instance of this prefix is scoped to a
specific context-group-version-kind, which the user interactively
selects via completion at instantiation time. `kele-resource` comes with
a single suffix at the moment, corresponding to the existing `kele-get`
functionality.

`kele-resource` also includes the ability to override context and
namespace via infix arguments. This functionality has rough edges and
known issues, which have been itemized in tickets for follow-up, e.g.
#117.