Implement "lazy" proxy manager (#153)
Closes #72.
Contributes to #151.

With this change, we decouple startup of a context's proxy server from
its "readiness" checks. The readiness check effectively becomes
"just-in-time," taking place right when a subroutine actually intends to
use the server. This prevents false-positive timeout errors from the
proxy server spinning up. This allows top-level commands, e.g. the
`kele-resource` prefix, to initialize the proxy server and allow it to
"preheat" before a suffix like `kele-get` actually gets around to using
it.