feature: suffix for restarting deployment (#188)
Closes #187.

This PR introduces the notion of kind-specific suffixes. A dedicated
collection of suffixes for `kele-resource` will show up depending on the
resource kind selected and whether or not special-case actions have been
defined on them.

This PR introduces the first such "special-case action" in Kele. Namely,
for Deployments, we build a suffix for restarting them, leveraging
`kubectl rollout restart` under the hood.