fix: handle uppercase letters in cluster URL
We use `url-generic-parse-url` to extract the host and port out of kubeconfig
cluster URLs. This has the unfortunate side effect (maybe due to RFC 3986
compliance?) of downcasing the hostname, which results in lookup failure when
then using that host to look up in `kele--global-discovery-cache`.

To fix, we switch to case-insensitive lookup for
`kele--get-resourec-lists-for-context`.