fix: singularize resource names where appropriate
Closes #196.

Currently, `kele-resource` uses the plural form of the resource name,
e.g. `pods`, everywhere. This was done primarily out of convenience but
obviously is grammatically inaccurate in some places, e.g. "get a single pods."

Here we introduce a mechanism for looking up the canonical singular form of any
given resource name from the discovery cache and apply it to the `kele-resource`
suffix labels.